Give up trying to make dispatch not capture colors from base16-shell.
If you use plain base16-xresources, the terminal colors are awful
but the base16 colors in vim work fine.

When I try to use the base16-shell colors in xresources, the terminal
colors are fine but the base16 colors in vim are screwed.

The only way terminal colors will work is with the shell script, but
then dispatch captures them, which was the whole point of this endeavor.

I could go back and remove the xresources colors, but I'm just going to
keep them so I don't forget and try this wild goose chase again in the
future.